Create 10.Program to Display Factors of a Number
In the program, a positive integer entered by the user is stored in num.The for loop is iterated until i <= num is false.

In each iteration, whether num is exactly divisible by i is checked. It is the condition for i to be a factor of num.

if (num % i == 0) {
            printf("%d ", i);
}
Then the value of i is incremented by 1.